fix: Allow narrowing a `Relationship` by its `target-type`
TypeScript has the concept of discriminating unions by a shared field name. For
example:

```TypeScript
declare var relations: Relationship[];
for (const r of relations) {
  switch (r["target-type"]) {
    case "area":
      r.area.name;
      break;
    case "release_group":
      r.release_group.title;
      break;
  }
}
```

On the `main` branch, this currently errors:

    error: TS18048 [ERROR]: 'r.area' is possibly 'undefined'.
          r.area.name;
          ~~~~~~
        at file:///home/michael/code/musicbrainz-ts/test/data/lookup.ts:1180:7

    TS18048 [ERROR]: 'r.release_group' is possibly 'undefined'.
          r.release_group.title;
          ~~~~~~~~~~~~~~~
        at file:///home/michael/code/musicbrainz-ts/test/data/lookup.ts:1183:7

    Found 2 errors.

With the changes in this commit, the same code type-checks. (It also blocks you
from accessing `r.release_group` under the `"area"` branch, or `r.area` under
the `"release_group"` branch, etc.).

The only way I could get this to work was by partially re-enabling the
distributivity that was disabled in 72e3ba55688793fc4e26d02d402b1763243a77fa; I
kept it disabled for `string` and `number`, which I believe are the only two
primitives *contained in arrays* we need to be concerned about.

A similar patch could be applied to collections.