feat: Autocomplete include parameters
Restrict type of include parameters from the fallback overload of the
lookup method to all possible values for all entity types.

This will also suggest include parameters which are not supported for
the currently selected entity type, but that is still better than no
having no suggestions at all (because of the fallback to string).

Unfortunately the autocomplete somehow falls back to suggesting
`AreaInclude` values for all `inc` values after the first one.