feat: Autocomplete all include parameters and make `lookup` generic
Replace all overloads of the `lookup` method with one generic interface.
Now that we have `EntityTypeMap` and `EntityIncludeMap` this approach no
longer causes "excessively deep type" errors.

The trick to make this work is to pass the generic parameter to the map
before accessing its properties, since accesssed properties of the map
can not be generic on their own.

Lots of imports and an unused helper type can now be dropped as well.