fix: Allow mixed relationship array assignments
Fixes the issue described in #4.

In `A extends any ? A[] : never`, if `A` is the union type `B | C`, then the
conditional type will evaluate to `B[] | C[]`.

In order to produce `(B | C)[]`, the array syntax `[]` must be lifted out of
the conditional type so that it's not distributed to each item in the union.

With respect to issue #4, we can achieve this by moving the problematic
conditional type expression into `WithIncludes`, which now acts on an unknown
`Data` type parameter:

 * If `Data` is an object, then `WithIncludes` behaves like before.
 * Otherwise, it now returns `Data` back.